Rename dfoo executables to dask-foo
Perform the following renaming

*  dscheduler -> dask-scheduler
*  dworker -> dask-worker
*  dcenter -> deprecated
*  dcluster -> dask-ssh

The old executables are still around but print a deprecation warning

Fixes #314
