+++ distributed/cluster.py
@@ -16,7 +16,7 @@ from threading import Thread
 from toolz import merge
 
 # These are handy for creating colorful terminal output to enhance readability
-# of the output generated by dcluster.
+# of the output generated by dask-ssh.
 class bcolors:
     HEADER = '\033[95m'
     OKBLUE = '\033[94m'
@@ -29,6 +29,7 @@ class bcolors:
 
 def async_ssh(cmd_dict):
     import paramiko
+    import paramiko.ssh_exception
     ssh = paramiko.SSHClient()
     ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
 
@@ -52,9 +53,10 @@ def async_ssh(cmd_dict):
             # Connection successful, break out of while loop
             break
 
-        except (paramiko.ssh_exception.SSHException, paramiko.ssh_exception.PasswordRequiredException) as e:
+        except (paramiko.ssh_exception.SSHException,
+                paramiko.ssh_exception.PasswordRequiredException) as e:
 
-            print('[ dcluster ] : ' + bcolors.FAIL +
+            print('[ dask-ssh ] : ' + bcolors.FAIL +
                   'SSH connection error when connecting to {addr}:{port} to run \'{cmd}\''.format(addr = cmd_dict['address'],
                                                                                                   port = cmd_dict['ssh_port'],
                                                                                                   cmd = cmd_dict['cmd']) +
@@ -69,7 +71,7 @@ def async_ssh(cmd_dict):
             # attempts to retry.
             retries += 1
             if retries >= 3:
-                print( '[ dcluster ] : ' + bcolors.FAIL + 'SSH connection failed after 3 retries. Exiting.' + bcolors.ENDC)
+                print( '[ dask-ssh ] : ' + bcolors.FAIL + 'SSH connection failed after 3 retries. Exiting.' + bcolors.ENDC)
 
                 # Connection failed after multiple attempts.  Terminate this thread.
                 os._exit(1)
@@ -150,12 +152,12 @@ def async_ssh(cmd_dict):
 
 
 def start_scheduler(logdir, addr, port, ssh_username, ssh_port, ssh_private_key):
-    cmd = '{python} -m distributed.cli.dscheduler --port {port}'.format(python=sys.executable, port=port, logdir=logdir)
+    cmd = '{python} -m distributed.cli.dask_scheduler --port {port}'.format(python=sys.executable, port=port, logdir=logdir)
 
     # Optionally re-direct stdout and stderr to a logfile
     if logdir is not None:
         cmd = 'mkdir -p {logdir} && ' + cmd
-        cmd += '&> {logdir}/dscheduler_{addr}:{port}.log'.format(addr=addr,
+        cmd += '&> {logdir}/dask_scheduler_{addr}:{port}.log'.format(addr=addr,
                 port=port, logdir=logdir)
 
     # Format output labels we can prepend to each line of output, and create
@@ -183,7 +185,7 @@ def start_scheduler(logdir, addr, port, ssh_username, ssh_port, ssh_private_key)
 def start_worker(logdir, scheduler_addr, scheduler_port, worker_addr, nthreads, nprocs,
                  ssh_username, ssh_port, ssh_private_key):
 
-    cmd = '{python} -m distributed.cli.dworker {scheduler_addr}:{scheduler_port} --host {worker_addr} --nthreads {nthreads} --nprocs {nprocs}'.format(
+    cmd = '{python} -m distributed.cli.dask_worker {scheduler_addr}:{scheduler_port} --host {worker_addr} --nthreads {nthreads} --nprocs {nprocs}'.format(
         python = sys.executable,
         scheduler_addr = scheduler_addr, scheduler_port = scheduler_port,
         worker_addr = worker_addr,
@@ -193,7 +195,7 @@ def start_worker(logdir, scheduler_addr, scheduler_port, worker_addr, nthreads, 
     # Optionally redirect stdout and stderr to a logfile
     if logdir is not None:
         cmd = 'mkdir -p {logdir} && ' + cmd
-        cmd += '&> {logdir}/dscheduler_{addr}.log'.format(
+        cmd += '&> {logdir}/dask_scheduler_{addr}.log'.format(
             addr = worker_addr, logdir = logdir)
 
     label = 'worker {addr}'.format(addr = worker_addr)
@@ -232,7 +234,7 @@ class Cluster(object):
         # Generate a universal timestamp to use for log files
         import datetime
         if logdir is not None:
-            logdir = os.path.join(logdir, "dcluster_" + datetime.datetime.now().strftime("%Y-%m-%d_%H:%M:%S"))
+            logdir = os.path.join(logdir, "dask-ssh_" + datetime.datetime.now().strftime("%Y-%m-%d_%H:%M:%S"))
             print(bcolors.WARNING + 'Output will be redirected to logfiles stored locally on individual woker nodes under "{logdir}".'.format(logdir=logdir) + bcolors.ENDC)
         self.logdir = logdir
 
@@ -282,4 +284,4 @@ class Cluster(object):
         return self
 
     def __exit__(self, *args):
-        self.close()
+        self.shutdown()
